requirements.txt
# Astro Runtime includes the following pre-installed providers packages:
# https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

apache-airflow-providers-dbt-cloud
astronomer-cosmos
astronomer-cosmos[google]
dbt-bigquery
pandas
pyarrow
google-cloud-storage
apache-airflow-providers-google 
pytest
